requirements.txt
################################################################################
# This requirements file has been automatically generated from `pip`.
#
################################################################################

flake8
pytest
pyyaml
python-dateutil
